+++ bricks/local_storage/CHANGELOG.md
@@ -1,3 +1,3 @@
 # 0.1.0+1
 
-- Added local storage repository brick
+- Added local storage service brick